// src/app/data/services/rest/model/user-update-schema.ts
/**
 * AngularCourse
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserUpdateSchema { 
    firstName?: string | null;
    lastName?: string | null;
    stack?: Array<string> | null;
    city?: string | null;
    description?: string | null;
}